Create ApiProcessorFactory class
-This factory class has inside the mehod createInstance which gives the
instance of ApiProcesssor in according to the category and to the
implementation. for example groceryCategory could be "fruits" or
"bakery", if we have an apiProcessor implementation for fruits category
or bakery, then we will get that instance, and we can call the
implemented methods inside this instance. In our case we are able to get
an implementation for "fruits" category which is provided by
ApiProcessorFruitImpl class. this class in fact implement the process
method of ApiProcesssor interface.