Create GroceryOrderController class
-This controller intercepts the request sent to Grocery. If the request
has also the fruits food category, then this will be handled by getFruit
method. this method call the run method of groceryProcessor and passes
to it the food category we need to process. If there is any
implementation of ApiProcesssor that can process the fruits category, we
will receive the response in the Json format.