Create GroceryRequestSender class
-This class contains fetchGroceryFruitItems method that is responsible
to fetch the fruits information. A call to buildResultJsonArray is
performed. we pass a productList to buildResultJsonArray. the result of
this call is a json array with all food information related to all
products in the html page. ProductList it's a Jsoup object which is an
implementation of array List of element. the elements we fetch from the
endpoint are all alements have productinfo as class name.